Fix testing
 - move common validation function into auxiliary and adjust testing
 functions accordingly
 - change set names in testing to new naming

Note: one of the tests has slightly different results now, presumably
due to the linearization of the new tests?
