+++ examples/2periods_new_build_2zones_singleBA/results/summary_results.txt
@@ -5,10 +5,9 @@
 --> New Generation Capacity <--
                               New Capacity (MW)
 load_zone technology  period                   
-Zone1     unspecified 2020                   24
-                      2030                    1
-Zone2     unspecified 2020                   28
-                      2030                    4
+Zone1     unspecified 2020                   25
+Zone2     unspecified 2020                   27
+                      2030                    6
 
 ### OPERATIONAL RESULTS ###
